Make `main` optional when compiling to WASM
Not sure this is idiomatic. However, I noticed, that when compiling to e.g. a yew app, which has a `main`, it conflicts with the existing `main` from `kord`.
The double negation of `not("wasm_no_main")` is because I didn't want to make this a breaking change, which I guess a feature `"wasm_main"` would have been.
Let me know if you prefer a feature `"wasm_main"` as part of the `default-features` perhaps, or something else entirely.